Started working on Settings Page
