--- Cheers/Controller/SettingsViewController.swift
@@ -1,35 +1,0 @@
-//
-//  SettingsViewController.swift
-//  Cheers
-//
-//  Created by Jordan Abushahla on 3/7/18.
-//  Copyright © 2018 University of San Diego. All rights reserved.
-//
-
-import UIKit
-
-class SettingsViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destinationViewController.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
